Only attempt to inject files if the injection disk exists
_inject_data was checking for the existence of the image to inject
into. However, the only caller of _inject_data has just created that
image. We move the _inject_data call up to where the image is created,
and pass the existing image object in directly. We modify _inject_data
to remove the existence check as, bar a delete race which was present
anyway, we know it exists.

As a bonus, this removes the last user of check_image_exists outside
imagebackend. This api is problematic as it almost always (including
this case) points to a race condition. The remaining callers can be
dealt with within a single module.

Implements: blueprint libvirt-instance-storage

Change-Id: Icb926e64151048c6ce8353de604417d5d79858e4