Remove fake_imagebackend.Raw and cleanup dependent tests
fake_imagebackend had a Raw class which was used by only a small
number of tests. This change refactors the tests to avoid the
requirement for this class by using mock. It also combines the common
code in the 2 rescue tests, as there was a large amount of
duplication.

Implements: blueprint libvirt-instance-storage

Change-Id: Iffdc0c66f1ec4428b98fffc807af636dd491a080