Make import_file do its own cleanup
import_file is currently used only by the Rbd backend, and only for
importing a config file. The interface requires the caller to cleanup
after it, which means that code outside the Rbd driver is doing:

  if CONF.libvirt.images_type == 'rbd':
      os.unlink(configdrive_path)

This change turns import_file into a context manager, and implements
it for the only current users: NoBacking and Rbd. This allows us to
handle the required cleanup in the Rbd backend rather than pushing it
to the caller.

Implements: blueprint libvirt-instance-storage

Change-Id: Ibc39fb2d5e3b7e1d5d1fc0386db2fe633e8897af