Rename Image.check_image_exists to Image.exists()
Apart from being shorter, this is also another step away from the
overloading of the term 'image' in the image backend. An 'image' in
this code refers either to the thing which glance stores, or an
instance's disk. Image in the latter context has come about because
that's what libvirt calls it. However, as it's overloaded in nova we
should never have used it. At some point I'd like to rename these all
to 'disk'.

Implements: blueprint libvirt-instance-storage

Change-Id: Ic15e8412971b2a3f8e87ec68605c4753e23f94b5