Introduce ImageCacheLocalPool
The code to download an image from glance or another compute host and
store it in the local image cache is currently variously distributed
around libvirt/driver and libvirt/imagebackend. This makes it very
hard to understand, fix, extend, and test.

This class implements all current behaviour relating to writing to the
image cache. This patch does not introduce a user of this code, but it
is used extensively in the series which follows.

Implements: blueprint libvirt-instance-storage

Change-Id: I4d6bef95bde872f9f369e1f0c1b6fc0f53242a5f