Introduce ImageCacheLocalDir
The code to download an image from glance or another compute host and
store it in the local image cache is currently variously distributed
around libvirt/driver and libvirt/imagebackend. This makes it very
hard to understand, fix, extend, and test.

This class implements all current behaviour relating to writing to the
image cache. This patch does not introduce a user of this code, but it
is used extensively in the series which follows.

Because we need to coordinate locking with ImageCacheManager, we
create 3 new common lock functions for use by both ImageCacheManager
and ImageCacheLocalDir, and update ImageCacheManager to use them. The
locks created are identical to those created previously, so existing
unit tests continue to validate them.

Implements: blueprint libvirt-instance-storage

Change-Id: I4d6bef95bde872f9f369e1f0c1b6fc0f53242a5f