Add Image.create_from_func
Image.cache() currently has different behaviour for 'generated' disks
and disks created from a glance image. This differing behaviour varies
from backend to backend. For example, it can be seen in the Lvm
backend, which does:

  generated = 'ephemeral_size' in kwargs

i.e. it requires this context but doesn't have it. It gets it by
grepping the arguments to the create function it was passed for an
ephemeral size, and infers the type based on that. The NoBacking
backend uses image_id instead, so it's not even consistent.

The interface is to pass a create function to Image.cache(), and
separately all of the arguments which will eventually be passed to it
if we need to call it.

This new interface makes it explicit to the backend that we are
creating a generated disk by having a separate method for doing this.
We also greatly simplify the calling interface by passing in a lambda
which takes a single argument: the destination path.

This patch also highlights a difference in caching behaviour between
backends: the Qcow2, Rdb, and Ploop backends currently cache function
output, whereas NoBacking and Lvm don't. This should eventually be made
consistent, but this patch explicitly aims to keep the current
behaviour.

Implements: blueprint libvirt-instance-storage

Change-Id: I2a8cc5f94a76d7db5f23109eccb175e66a8436d7