Rename Raw backend to NoBacking
As mentioned in a comment (which this patch deletes), calling this
class 'Raw' was confusing, because it is not always raw. It is also a
source of bugs, because some code assumes that the format is always
raw, which it is not. This patch does not fix those bugs.

We rename it to 'NoBacking', which describes it accurately. We also add
doctext describing what it does.

DocImpact

The config option libvirt.images_type gets an additional value:
'nobacking'. The effect of this is identical to setting the value
'raw'.

Implements: blueprint libvirt-instance-storage

Change-Id: I93f0a2cc568b60c2b3f7509449167f03c3f30fb5